requirements.txt
Django 
Pillow 
djangorestframework 
djangorestframework-simplejwt 
django-cors-headers 
pywin32 
whitenoise 